Add adv_train_type to the output file name

+++ train_protonet_he.py
@@ -77,7 +77,7 @@ evaluation_taskloader = DataLoader(
 # Model #
 #########
 model = get_few_shot_he_encoder(num_input_channels, final_layer_size)
-model.to(device, dtype=torch.double)
+model.to(device, dtype=torch.float)
 
 ############
 # Training #
